Refractor code from based-module to module-style
Get data by indentity id
