+++ src/main/java/com/dawn/jat/illuminati/comment/controller/CommentController.java
@@ -1,0 +1,4 @@
+package com.dawn.jat.illuminati.comment.controller;
+
+public class CommentController {
+}
