+++ src/main/java/com/dawn/jat/illuminati/comment/entity/CommentEntity.java
@@ -1,0 +1,4 @@
+package com.dawn.jat.illuminati.comment.entity;
+
+public class CommentEntity {
+}
