+++ src/main/java/com/dawn/jat/illuminati/comment/repository/CommentRepository.java
@@ -1,0 +1,4 @@
+package com.dawn.jat.illuminati.comment.repository;
+
+public class CommentRepository {
+}
