+++ src/main/java/com/dawn/jat/illuminati/comment/service/CommentService.java
@@ -1,0 +1,4 @@
+package com.dawn.jat.illuminati.comment.service;
+
+public class CommentService {
+}
