--- src/main/java/com/dawn/jat/illuminati/entities/Post.java
@@ -1,62 +1,0 @@
-package com.dawn.jat.illuminati.entities;
-
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-@Document(collection = "post")
-public class Post {
-
-    @Id
-    private String id;
-    private String title;
-    private String tag;
-    private String brief;
-    private String time;
-
-    @Override
-    public String toString() {
-        return "Post [id=" + id + ", title=" + title + ", tag=" + tag + ", brief=" + brief
-                + ", time=" + time + "]";
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getTag() {
-        return tag;
-    }
-
-    public void setTag(String tag) {
-        this.tag = tag;
-    }
-
-    public String getBrief() {
-        return brief;
-    }
-
-    public void setBrief(String brief) {
-        this.brief = brief;
-    }
-
-    public String getTime() {
-        return time;
-    }
-
-    public void setTime(String time) {
-        this.time = time;
-    }
-
-}
