+++ src/main/java/com/dawn/jat/illuminati/post/repository/PostRepository.java
@@ -1,0 +1,9 @@
+package com.dawn.jat.illuminati.post.repository;
+
+import com.dawn.jat.illuminati.post.entity.PostEntity;
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PostRepository extends MongoRepository<PostEntity, String> {
+}
