+++ src/main/resources/application-prod.properties
@@ -1,0 +1,2 @@
+app.name=Illuminati on Production
+app.description=${app.name}
