+++ src/main/resources/application-staging.properties
@@ -1,0 +1,2 @@
+app.name=Illuminati on Staging
+app.description=${app.name}
