+++ src/main/resources/application.properties
@@ -1,2 +1,6 @@
+app.name=Illuminati on Local
+app.description=${app.name}
+spring.profiles.active=${SPRING_ACTIVE_PROFILE:"dev"}
+
 spring.data.mongodb.uri=mongodb://localhost:27017
 spring.data.mongodb.database=sample_post
