+++ src/test/java/com/dawn/jat/illuminati/PostControllerTest.java
@@ -1,0 +1,4 @@
+package com.dawn.jat.illuminati;
+
+public class PostControllerTest {
+}
