+++ src/test/java/com/dawn/jat/illuminati/hello/HelloControllerTest.java
@@ -2,6 +2,7 @@ package com.dawn.jat.illuminati.hello;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import com.dawn.jat.illuminati.hello.controller.HelloController;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
