+++ src/test/java/com/dawn/jat/illuminati/integration/post/PostEntityControllerTest.java
@@ -2,7 +2,7 @@ package com.dawn.jat.illuminati.integration.post;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.dawn.jat.illuminati.post.PostController;
+import com.dawn.jat.illuminati.post.controller.PostController;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -10,7 +10,7 @@ import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.boot.web.server.LocalServerPort;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class PostControllerTest {
+public class PostEntityControllerTest {
     @Autowired
     private PostController controller;
 
@@ -26,13 +26,13 @@ public class PostControllerTest {
     private TestRestTemplate restTemplate;
 
     @Test
-    public void getPostByAvaiTitleShouldRespOK() throws Exception {
-        assertThat(this.restTemplate.getForObject("http://localhost:" + port + "/post/Java",
+    public void getPostByAvaiIdShouldRespOK() throws Exception {
+        assertThat(this.restTemplate.getForObject("http://localhost:" + port + "/post/1",
                 String.class)).contains("OK");
     }
 
-    public void getPostByUnavaiTitleShouldRepBadReq() throws Exception {
-        assertThat(this.restTemplate.getForObject("http://localhost:" + port + "/post/Ruby",
+    public void getPostByUnavaiIdShouldRepBadReq() throws Exception {
+        assertThat(this.restTemplate.getForObject("http://localhost:" + port + "/post/Java",
                 String.class)).contains("Cannot find post");
     }
 }
